+++ app/page/authorization/transaction.tsx
@@ -1,35 +1,22 @@
 import * as ethers from "ethers"
 import { useEffect, useState } from "react"
-import { useClsState } from "use-cls-state"
 import { useHashLocation } from "wouter/use-hash-location"
 
 import { useProviderContext } from "~app/context/provider"
 import { Path } from "~app/path"
-import {
-  useAccount,
-  useAction,
-  useNetwork,
-  usePendingTransactions
-} from "~app/storage"
+import { useAccount, useNetwork, usePendingTransactions } from "~app/storage"
 import type { Account } from "~packages/account"
-import {
-  UserOperationV0_6,
-  UserOperationV0_7,
-  type UserOperation
-} from "~packages/bundler/userOperation"
 import type { Paymaster } from "~packages/paymaster"
 import { NullPaymaster } from "~packages/paymaster/NullPaymaster"
 import { VerifyingPaymaster } from "~packages/paymaster/VerifyingPaymaster"
 import { ETH } from "~packages/token"
-import { WaalletRpcMethod } from "~packages/waallet/rpc"
 import { AccountStorageManager } from "~storage/local/manager"
-import {
-  TransactionType,
-  type Network,
-  type TransactionPending
-} from "~storage/local/state"
+import { type Network, type TransactionPending } from "~storage/local/state"
 
-type PaymentOption = {
+import { usePayment } from "./usePayment"
+import { useUserOperation } from "./useUserOperation"
+
+export type PaymentOption = {
   name: string
   paymaster: Paymaster
 }
@@ -80,9 +67,6 @@ function UserOperationConfirmation(props: {
   sender: Account
   network: Network
 }) {
-  const { tx, sender, network } = props
-
-  const [, navigate] = useHashLocation()
   const { provider } = useProviderContext()
 
   const paymentOptions: PaymentOption[] = [
@@ -102,138 +86,19 @@ function UserOperationConfirmation(props: {
       })
     }
   ]
-  const {
-    getERC4337TransactionType,
-    markERC4337TransactionSent,
-    markERC4337TransactionRejected
-  } = useAction()
-
-  /* Payment */
 
   const [paymentOption, setPaymentOption] = useState<PaymentOption>(
     paymentOptions[0]
   )
-  const [payment, setPayment] = useState({
-    token: ETH,
-    tokenFee: 0n
-  })
-  const [paymentCalculating, setPaymentCalculating] = useState(false)
+  const {
+    userOp,
+    userOpResolving,
+    sendUserOperation,
+    rejectUserOperation,
+    userOpEstimating
+  } = useUserOperation({ ...props, paymentOption })
 
-  /* User Operation */
-
-  const [userOp, setUserOp] = useClsState<UserOperation>(null)
-  const [userOpResolving, setUserOpResolving] = useState(false)
-  const [userOpEstimating, setUserOpEstimating] = useState(false)
-
-  const sendUserOperation = async () => {
-    setUserOpResolving(true)
-    try {
-      const entryPoint = await sender.getEntryPoint()
-
-      userOp.setPaymasterAndData(
-        await paymentOption.paymaster.requestPaymasterAndData(userOp)
-      )
-      userOp.setSignature(
-        await sender.sign(userOp.hash(entryPoint, network.chainId))
-      )
-      const userOpHash = await provider.send(
-        WaalletRpcMethod.eth_sendUserOperation,
-        [userOp.unwrap(), entryPoint]
-      )
-      if (!userOpHash) {
-        throw new Error("Fail to send user operation")
-      }
-      await markERC4337TransactionSent(tx.id, {
-        entryPoint,
-        userOp,
-        userOpHash
-      })
-    } catch (e) {
-      // TOOD: Show error on page
-      console.error(e)
-      setUserOpResolving(false)
-    }
-  }
-
-  const rejectUserOperation = async () => {
-    setUserOpResolving(true)
-    try {
-      await markERC4337TransactionRejected(tx.id, {
-        entryPoint: await sender.getEntryPoint(),
-        userOp
-      })
-      navigate(Path.Index)
-    } catch (e) {
-      console.error(e)
-      setUserOpResolving(false)
-    }
-  }
-
-  const estimateGas = async (userOp: UserOperation) => {
-    const paymasterAndData =
-      await paymentOption.paymaster.requestPaymasterAndData(userOp, true)
-    userOp.setPaymasterAndData(paymasterAndData)
-
-    const gasFee = await provider.send(
-      WaalletRpcMethod.custom_estimateGasPrice,
-      []
-    )
-    userOp.setGasFee(gasFee)
-
-    const gasLimit = await provider.send(
-      WaalletRpcMethod.eth_estimateUserOperationGas,
-      [userOp.unwrap(), await sender.getEntryPoint()]
-    )
-    userOp.setGasLimit(gasLimit)
-  }
-
-  useEffect(() => {
-    async function setupUserOp() {
-      const transactionType = getERC4337TransactionType(
-        tx.networkId,
-        await sender.getEntryPoint()
-      )
-      const execution = await sender.buildExecution(tx)
-      const userOp =
-        transactionType === TransactionType.ERC4337V0_6
-          ? UserOperationV0_6.wrap(execution)
-          : UserOperationV0_7.wrap(execution)
-      await estimateGas(userOp)
-      setUserOp(userOp)
-    }
-    setupUserOp()
-  }, [])
-
-  useEffect(() => {
-    async function estimateUserOp() {
-      setUserOpEstimating(true)
-      await estimateGas(userOp)
-      setUserOp(userOp)
-      setUserOpEstimating(false)
-    }
-    if (!userOp) {
-      return
-    }
-    estimateUserOp()
-  }, [paymentOption])
-
-  useEffect(() => {
-    async function calculatePayment() {
-      setPaymentCalculating(true)
-      setPayment({
-        ...payment,
-        tokenFee: await paymentOption.paymaster.quoteFee(
-          userOp.calculateGasFee(),
-          ETH
-        )
-      })
-      setPaymentCalculating(false)
-    }
-    if (!userOp) {
-      return
-    }
-    calculatePayment()
-  }, [JSON.stringify(userOp?.unwrap())])
+  const { payment, paymentCalculating } = usePayment({ userOp, paymentOption })
 
   if (!userOp) {
     return <></>
